authmate: Fix JSON tag for session container field
NeoFS API uses lower camel case for JSON field names. Verb tag is
notated in lower camel case, so it makes completely no sense to declare
container tag in the upper one. According to the struct definition and
`encoding/json` parser, the change is fully invisible: both
`containerID` and `ContainerID` JSON fields are parsed the same
(extended unit test shows this). However, it's still better to follow
the recommended format and unite it for all fields.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>