cache: rework ACL state cache into a generic EACL cache
It's much more useful this way especially if we have many requests with EACLs.
ACL status can be deduced from EACL, so we don't need to cache it separately
(at least for now, walking through the list takes some time, but it's nothing
like getting EACL from the network that we did previously in many cases).

Signed-off-by: Roman Khimov <roman@nspcc.ru>