tests: Stabilize test
Closes #1130.

Sometimes, the runtime took a bit more time, and data was wiped out from the cache.
Therefore, sometimes we get nil from the condition. We reduced the condition check
for 1 second to ensure we had no nil result for the Never check. Simultaneously,
we increased Eventually check interval. This condition checks that the cache is going to be empty
during these two seconds. In one second and a few milliseconds in fact.

Signed-off-by: Evgenii Baidakov <evgenii@nspcc.io>